[UPDATE] Added comments.

+++ reddit/models/comment.py
@@ -1,10 +1,13 @@
 from reddit.extensions import db
 from typing import Optional
 
+class PermissionError(Exception):
+    pass
 
 class Comment:
     def __init__(self, cid: Optional[int] = None):
         self.id = cid
+        self.content = None
 
     @staticmethod
     def add(content: str, postid: int, authorid: int, parentid: int):
@@ -21,14 +24,25 @@ class Comment:
                 )
             )
 
-    def delete(self):
+    def delete(self, userId: int):
         with db as cursor:
             cursor.execute(
-                "DELETE FROM Comments WHERE Id = ?;", (self.id)
+                "DELETE FROM Comments WHERE Id = ? AND AuthorId = ?;", (self.id, userId)
             )
+            if cursor.rowcount == 0:
+                raise PermissionError()
 
     def edit(self, content: str):
         with db as cursor:
             cursor.execute(
-                "UPDATE Comments SET Content = ?;", (content)
+                "UPDATE Comments SET Content = ?;", (content,)
             )
+
+    def fetch(self):
+        with db as cursor:
+            cursor.execute(
+              "SELECT Content FROM Comments WHERE id = ?;",(self.id,)  
+            )
+
+            row = cursor.fetchone()
+            self.content = row[0]
